Fix use of _flush callback
The _flush callback can't be called until the stream is in a final
state.  This is when the inflater has either ended or errored.  Its
.end() callback is only called on 'finish', which is neither of those.
Instead call the callback on the 'end' or 'error' event, but be sure not
to forward any error, which is instead emitted by the event forwarding
code.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>