Generalize function names a bit
The "inflate"/"inflater" terminology was both overly specific and
confusing/misleading due to confusion between the inflate format,
Inflate class, and the verb for decompression.  Switch to the term
format when referring to the data format and class used to decode it,
and decoder for the instance which does the decoding.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>