feat: create custom Error constructors
As a result of nodejs/node#33857, Node errors are no longer custom error
classes and .constructor is a built-in Error constructor.  Therefore, to
support Node 15 and later, define Error constructors for the codes used
by this module in this module, as done by readable-stream and other
projects.

Note: Much of the code is copied from lib/internal/errors.js @ v15.0.1.
ESLint rules are relaxed to make it easier to minimize divergence and
make updates easier.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>