Add exceptions for autoDestroy tests
The zlib autoDestroy behavior changed in Node.js v14.  The change is
positive (in my opinion) and this module has adopted the behavior with
one exception:  `autoDestroy: false` is not respected for errors from
the zlib bindings.  Mimicking this behavior would be difficult, since
the 'error' event handler would need to determine whether the error was
caused by bad data, inconsistent (since autoDestroy is respected if
'error' is emitted for other reasons), and cause significant divergence
from zlib on previous Node.js versions unless the behavior were
version-dependent (which I generally prefer to avoid).

Ignore 'close' in these tests on Node.js v14 and v15 for now.  Once v16
is released, if the behavior is still present, reconsider.

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>