.github: check-latest for actions/setup-node@v2
By default, if a cached version matches the node-version option,
actions/setup-node@v2 will use it.  Since `*` matches everything, the
latest cached version (currently 14.17.4) is used for jobs with version
`*`.  This is not ideal, since we want to test with the earliest and
latest versions of Node.js, and v16 has been available since 2021-04-20.
Therefore, set check-latest: true for `*` versions.

Note: check-latest is set false for `12.20`, since any patch release of
12.20 is fine for our needs (although it is not expected to be in the
cache).

Signed-off-by: Kevin Locke <kevin@kevinlocke.name>